Re-organise modules to hide internals from public_api_tests.rs
After moving the public API tests from the integrations `/tests`
folder to `/src/public_api_tests.rs` in order to be able to re-use the
test harness code between internal and public API tests, we lost the
ability to hide the internals from these tests, which would ensure
that we truly test the public API of `thor`.

Moving the channel APIs from `/src/lib.rs` to `/src/channel.rs`
allows us to regain the ability to hide the internals from the public
API tests. To do this, we have to move all the protocols under the
`channel` module and make use of `pub(in path)` notation.
