--- thor/src/protocols.rs
@@ -1,5 +1,0 @@
-pub mod close;
-pub mod create;
-pub mod punish;
-pub mod splice;
-pub mod update;
