+++ thor/src/public_api_tests.rs
@@ -107,97 +107,6 @@ fn channel_update() {
 }
 
 #[test]
-fn punish_publication_of_revoked_commit_transaction() {
-    let mut runtime = build_runtime();
-
-    let tc_client = testcontainers::clients::Cli::default();
-    let bitcoind = Bitcoind::new(&tc_client, "0.19.1").unwrap();
-    runtime.block_on(bitcoind.init(5)).unwrap();
-
-    let fund_amount_alice = Amount::ONE_BTC;
-    let fund_amount_bob = Amount::ONE_BTC;
-    let (balance_alice, balance_bob) = generate_balances(fund_amount_alice, fund_amount_bob);
-
-    let time_lock = 1;
-
-    let (alice_wallet, bob_wallet) = runtime
-        .block_on(make_wallets(&bitcoind, fund_amount_alice, fund_amount_bob))
-        .unwrap();
-    let (mut alice_transport, mut bob_transport) = make_transports();
-
-    let alice_create = Channel::create(
-        &mut alice_transport,
-        &alice_wallet,
-        balance_alice,
-        time_lock,
-    );
-    let bob_create = Channel::create(&mut bob_transport, &bob_wallet, balance_bob, time_lock);
-
-    let (mut alice_channel, mut bob_channel) = runtime
-        .block_on(futures::future::try_join(alice_create, bob_create))
-        .unwrap();
-
-    let after_open_balance_alice = runtime.block_on(alice_wallet.0.balance()).unwrap();
-    let after_open_balance_bob = runtime.block_on(bob_wallet.0.balance()).unwrap();
-
-    // Parties agree on a new channel balance: Alice pays 0.5 a Bitcoin to Bob
-    let payment = Amount::from_btc(0.5).unwrap();
-    let alice_balance = fund_amount_alice - payment;
-    let bob_balance = fund_amount_bob + payment;
-
-    let alice_update = alice_channel.update_balance(
-        &mut alice_transport,
-        Balance {
-            ours: alice_balance,
-            theirs: bob_balance,
-        },
-        time_lock,
-    );
-    let bob_update = bob_channel.update_balance(
-        &mut bob_transport,
-        Balance {
-            ours: bob_balance,
-            theirs: alice_balance,
-        },
-        time_lock,
-    );
-
-    runtime
-        .block_on(futures::future::try_join(alice_update, bob_update))
-        .unwrap();
-
-    // Alice attempts to cheat by publishing a revoked commit transaction
-
-    let signed_revoked_TX_c = alice_channel.latest_revoked_signed_TX_c().unwrap().unwrap();
-    runtime
-        .block_on(
-            alice_wallet
-                .0
-                .send_raw_transaction(signed_revoked_TX_c.clone()),
-        )
-        .unwrap();
-
-    // Bob sees the transaction and punishes Alice
-
-    runtime
-        .block_on(bob_channel.punish(&bob_wallet, signed_revoked_TX_c))
-        .unwrap();
-
-    let after_punish_balance_alice = runtime.block_on(alice_wallet.0.balance()).unwrap();
-    let after_punish_balance_bob = runtime.block_on(bob_wallet.0.balance()).unwrap();
-
-    assert_eq!(
-        after_punish_balance_alice, after_open_balance_alice,
-        "Alice should get no money back after being punished"
-    );
-    assert_eq!(
-        after_punish_balance_bob,
-        after_open_balance_bob + fund_amount_bob * 2 - Amount::from_sat(crate::TX_FEE) * 2,
-        "Bob should get all the money back after punishing Alice"
-    );
-}
-
-#[test]
 fn channel_collaborative_close() {
     let mut runtime = build_runtime();
